Refactor notification prefix logic by introducing reusable `applyPrefix` function and streamlining its usage across notification utilities.
